Refactor: Split vrt.spec.ts into Component-Specific Files
This commit refactors the monolithic `vrt.spec.ts` into smaller, component-specific visual regression test files. This improves maintainability and aligns with the goals of the QA epic.

Key changes:
- Created new VRT files for `TimerControls`, `SpotifyControls`, `Dashboard`, `ControlPanel`, `MockHRMClient`, and `HR-Related Components`.
- Migrated tests from `vrt.spec.ts` to the new files.
- Deleted the obsolete `vrt.spec.ts` and its corresponding snapshot directory.
- Fixed numerous issues in the VRT tests that were uncovered during the refactoring, including:
    - Corrected the dashboard route in the test setup to point to `/client/experimental`.
    - Added missing `data-testid` attributes to the `MockHRMClient` and `ExperimentalAnalyticsPage` components to allow for proper element selection.
    - Removed tests for components that no longer exist on the experimental dashboard (`TimerDisplay`, `HRTileGrid`).
    - Stabilized the dashboard test by removing unnecessary waits and ensuring the correct page is loaded.
- Updated snapshots for all new and modified tests.
