+++ app/client/experimental/components/ExperimentalAnalyticsPage.tsx
@@ -195,7 +195,7 @@ const ExperimentalAnalyticsPage = () => {
   }, [activeSession, totalCaloriesBurned])
 
   return (
-    <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
+    <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }} data-testid="dashboard">
       {view === 'active' && (
         <>
           <Box sx={{ mb: 3, display: 'flex', gap: 2 }}>
