+++ tests/playwright/vrt-hr-components.spec.ts
@@ -1,0 +1,59 @@
+
+import { type BrowserContext, type Page } from '@playwright/test'
+import { expect, test } from './fixtures'
+import {
+  getDynamicContentMasks,
+  getHrMasks,
+  setupVisualRegressionTest,
+} from './test-helpers'
+import { takeScreenshot } from './lib/visual'
+import { WAIT_TIMEOUTS, waitForPageReady } from './lib/waits'
+
+// Test suite configuration
+test.describe.configure({ mode: 'serial' })
+
+// Reusable page objects
+let controlPage: Page
+let dashboardPage: Page
+let mockPage: Page
+let context: BrowserContext
+
+// Test suite for VRT
+test.describe('Visual Regression Tests', () => {
+  // Centralized setup hook
+  test.beforeAll(async ({ browser }) => {
+    const setup = await setupVisualRegressionTest(browser)
+    context = setup.context
+    controlPage = setup.controlPage
+    dashboardPage = setup.dashboardPage
+    mockPage = setup.mockPage
+  })
+
+  // Centralized cleanup hook
+  test.afterAll(async () => {
+    await context?.close()
+  })
+
+  // Add a beforeEach hook to wait for the page to be ready before each test
+  test.beforeEach(async () => {
+    await waitForPageReady(controlPage)
+    await waitForPageReady(dashboardPage)
+  })
+
+  test.describe('HR-Related Components', () => {
+    test('dashboard with HR data', async () => {
+      await mockPage.getByLabel('Current BPM').fill('155')
+      await mockPage.getByRole('button', { name: 'Zone 4' }).click()
+
+      const dashboard = dashboardPage.getByTestId('dashboard')
+
+      await takeScreenshot(dashboard, 'dashboard-with-hr-data.png', {
+        maxDiffPixelRatio: 0.04,
+        mask: [
+          ...getDynamicContentMasks(dashboardPage),
+          ...getHrMasks(dashboardPage),
+        ],
+      })
+    })
+  })
+})
